# Remove redundant array copy for TiffFileWSIReader