@drbeh 2023-03-01
```python
from monai.data.wsi_reader import WSIReader
from monai.data.image_reader import WSIReader as WSIReader2

filename = "temp_CMU-1.tiff"
reader1 = WSIReader(backend="TiffFile")
reader2 = WSIReader2(backend="TiffFile")
obj1 = reader1.read(filename)
obj2 = reader2.read(filename)
```
```python
%%timeit
reader1.get_data(obj1, level=1)
```
current performance: 2.32 s ± 97.7
after fixing the issue: 1.45 s ± 12.9 ms

```python
%%timeit
reader2.get_data(obj2, level=1)
```
current performance:  1.97 s ± 50.2 ms
after fixing the issue:. 1.93 s ± 141 ms