@drbeh 2023-03-01
> thanks, is it possible to add a test to avoid regression?

Since it was a performance regression, I am not sure how we should set the criteria for the test as it is hardware dependent, especially when dealing with variability across different run of the same class. Do you have any suggestion? Maybe we should just check for regression whenever we have a major change.