@wyli 2023-03-01
Maybe just check theres no array copy, which means changing the patch would also change the input array?