@drbeh 2023-03-01
> Maybe just check theres no array copy, which means changing the patch would also change the input array?

Right but in this case there is mo use for the input array since the array input is loaded from disk and is being prepared based on the arguments, so if we copy the original array, it will be discarded anyway.