@wyli 2023-03-02
Sure, there was a similar performance bug earlier https://github.com/Project-MONAI/MONAI/issues/3322 hope we can add some tests to avoid reoccurring